requirements.txt
apache-airflow
matplotlib
seaborn
pytest
omegaconf
feast
great_expectations
dvc
hydra-core
pytest
pandas
scikit-learn
pendulum
shyaml
mlflow
giskard
evidently
Flask
flask-session2
psycopg2-binary
pyyaml
gradio
zenml
